moved walk forward base class
