+++ macrosynergy/learning/splitters/__init__.py
@@ -1,6 +1,7 @@
-from .base_panel_split import BasePanelSplit, KFoldPanelSplit
+from .base_panel_split import BasePanelSplit, KFoldPanelSplit, WalkForwardPanelSplit
 
 __all__ = [
     'BasePanelSplit',
     'KFoldPanelSplit',
+    'WalkForwardPanelSplit',
 ]
