+++ macrosynergy/learning/splitters/walk_forward_splitters.py
@@ -8,150 +8,9 @@ from typing import Optional, List
 import numpy as np
 import pandas as pd
 
-from base_panel_split import BasePanelSplit
+from base_panel_split import BasePanelSplit, WalkForwardPanelSplit
 
 from abc import ABC
-
-
-class WalkForwardPanelSplit(BasePanelSplit, ABC):
-    """
-    Base class for a generic walk-forward panel cross-validator.
-
-    Provides train/test indices to split a panel into train/test sets. Following an
-    initial training set construction, a forward test set is created. The training and
-    test set pair evolves over time by walking forward through the panel.
-
-    Parameters
-    ----------
-    min_cids : int
-        Minimum number of cross-sections required for the first training set.
-        Either start_date or (min_cids, min_periods) must be provided.
-        If both are provided, start_date takes precedence.
-    min_periods : int
-        Minimum number of time periods required for the first training set. Either
-        start_date or (min_cids, min_periods) must be provided. If both are
-        provided, start_date takes precedence.
-    start_date : Optional[str]
-        The targeted final date in the initial training set in ISO 8601 format.
-        Default is None. Either start_date or (min_cids, min_periods) must be provided.
-        If both are provided, start_date takes precedence.
-    max_periods : Optional[int]
-        The maximum number of time periods in each training set. If the maximum is
-        exceeded, the earliest periods are cut off. This effectively creates rolling
-        training sets. Default is None.
-    """
-
-    def __init__(
-        self,
-        min_cids,
-        min_periods,
-        start_date=None,
-        max_periods=None,
-    ):
-        # Checks
-        self._check_wf_params(
-            min_cids=min_cids,
-            min_periods=min_periods,
-            start_date=start_date,
-            max_periods=max_periods,
-        )
-
-        # Attributes
-        self.min_cids = min_cids
-        self.min_periods = min_periods
-        self.start_date = pd.Timestamp(start_date) if start_date else None
-        self.max_periods = max_periods
-
-    def _check_wf_params(
-        self, min_cids, min_periods, start_date, max_periods
-    ):
-        """
-        Type and value checks for the class initialisation parameters.
-
-        Parameters
-        ----------
-        min_cids : int
-            Minimum number of cross-sections required for the first training set.
-        min_periods : int
-            Minimum number of time periods required for the first training set.
-        start_date : Optional[str]
-            The targeted final date in the initial training set in ISO 8601 format.
-        max_periods : Optional[int]
-            The maximum number of time periods in each training set.
-        """
-        # min_cids
-        if not isinstance(min_cids, int):
-            raise TypeError(f"min_cids must be an integer. Got {type(min_cids)}.")
-        if min_cids < 1:
-            raise ValueError(
-                f"min_cids must be an integer greater than 0. Got {min_cids}."
-            )
-        # min_periods
-        if not isinstance(min_periods, int):
-            raise TypeError(f"min_periods must be an integer. Got {type(min_periods)}.")
-        if min_periods < 1:
-            raise ValueError(
-                f"min_periods must be an integer greater than 0. Got {min_periods}."
-            )
-        # start_date
-        if start_date is not None and not isinstance(start_date, str):
-            raise TypeError(f"start_date must be a string. Got {type(start_date)}.")
-        if start_date is not None:
-            try:
-                datetime.datetime.fromisoformat(start_date)
-            except ValueError:
-                raise ValueError(
-                    f"start_date must be in ISO 8601 format. Got {start_date}."
-                )
-        # max_periods
-        if max_periods is not None and not isinstance(max_periods, int):
-            raise TypeError(f"max_periods must be an integer. Got {type(max_periods)}.")
-        if max_periods is not None and max_periods < 1:
-            raise ValueError(
-                f"max_periods must be an integer greater than 0. Got {max_periods}."
-            )
-        
-    def _check_split_params(self, X, y, groups):
-        """
-        Type and value checks for the `split()` method parameters.
-
-        Parameters
-        ----------
-        X : pd.DataFrame
-            Pandas dataframe of features, multi-indexed by (cross-section, date). The
-            dates must be in datetime format. Otherwise the dataframe must be in wide
-            format: each feature is a column.
-
-        y : Union[pd.DataFrame, pd.Series]
-            Pandas dataframe or series of a target variable, multi-indexed by
-            (cross-section, date). The dates must be in datetime format. If a dataframe
-            is provided, the target variable must be the sole column.
-
-        groups : None
-            Ignored. Exists for compatibility with scikit-learn.
-        """
-        # X
-        if not isinstance(X.index, pd.MultiIndex):
-            raise ValueError("X must be multi-indexed.")
-        if not pd.api.types.is_datetime64_any_dtype(X.index.get_level_values(1)):
-            raise ValueError(
-                f"The dates in X must be datetime objects. Got {X.index.get_level_values(1).dtype} instead."
-            )
-        # y
-        if not isinstance(y.index, pd.MultiIndex):
-            raise ValueError("y must be multi-indexed.")
-        if not pd.api.types.is_datetime64_any_dtype(y.index.get_level_values(1)):
-            raise ValueError(
-                f"The dates in y must be datetime objects. Got {y.index.get_level_values(1).dtype} instead."
-            )
-        if not X.index.equals(y.index):
-            raise ValueError(
-                "The indices of the input dataframe X and the output dataframe y don't"
-                "match."
-            )
-        # groups
-        if groups is not None:
-            raise ValueError("groups is not supported by this splitter.")
 
 class ExpandingIncrementPanelSplit(WalkForwardPanelSplit):
     """
